[OT-98] [REFACTOR]: 폴더 정리에 따른 alias 1차 변경
