+++ src/app/(service)/mypage/profile/page.tsx
@@ -1,9 +1,9 @@
 // /mypage/profile 도메인
 import Header from "@/components/common/Header";
 import BackButton from "@/components/common/BackButton";
-import ProfileEditor from "@/components/mypage/profile/ProfileEditor";
-import EditFavoriteTagsUI from "@/components/mypage/profile/EditFavoriteTagsUI";
-import AccountActionButtons from "@/components/mypage/profile/AccountActionButtons";
+import ProfileEditor from "@/domains/mypage/components/profile/ProfileEditor";
+import EditFavoriteTagsUI from "@/domains/mypage/components/profile/EditFavoriteTagsUI";
+import AccountActionButtons from "@/domains/mypage/components/profile/AccountActionButtons";
 
 export default function MyPageProfile() {
   return (
