+++ src/domains/mypage/components/dashboard/TagStatsModalList.tsx
@@ -3,8 +3,8 @@
 import { useEffect, useRef, useState } from "react";
 import Image from "next/image";
 import { RecommendedContent } from "@/domains/mypage/types/dashboard";
-import LeftListScroll from "@/components/mypage/LeftListScroll";
-import RightListScroll from "@/components/mypage/RightListScroll";
+import LeftListScroll from "@/domains/mypage/components/LeftListScroll";
+import RightListScroll from "@/domains/mypage/components/RightListScroll";
 
 interface TagStatsModalListProps {
   items: RecommendedContent[];
